Fixed filepath conflicts
